move out criticality search tests
